Authorization and authentication complete, handling errores completed

+++ MEAN-Stack-Course/mean-course/backend/routes/posts.js
@@ -30,7 +30,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req,res,
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {'image'
     res.status(201).json({
@@ -40,6 +41,11 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req,res,
         id: createdPost._id
       }
     });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: 'Creating a post failed'
+    })
   });
 
 });
@@ -50,17 +56,25 @@ router.put("/:id", checkAuth,multer({storage: storage}).single("image"), (req,re
     const url = req.protocol + '://' + req.get("host");
     imagePath = url + "/images/" + req.file.filename;
   }
-  console.log(req.file);
   const post = new Post({
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Update successful"});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if(result.nModified > 0) {
+      res.status(200).json({message: "Update successful"});
+    } else {
+      res.status(401).json({message: "Not authorized"});
+    }
   })
+  .catch(erro => {
+    res.status(500).json({
+      message: "Couldn't update post"
+    })
+  });
 })
 
 router.get("",(req, res, next) => {
@@ -82,6 +96,11 @@ router.get("",(req, res, next) => {
         posts: fetchedPost,
         maxPosts: count
       });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: 'Fetching posts failed'
+      })
     });
 });
 
@@ -92,14 +111,26 @@ router.get('/:id', (req,res,next) => {
     } else {
       res.status(404).json({message: 'Post not found'});
     }
+  }).catch(error => {
+    res.status(500).json({
+      message: 'Fetching posts failed'
+    })
   });
 })
 
 router.delete("/:id", checkAuth,(req,res,next) => {
-   Post.deleteOne({_id: req.params.id}).then(result => {
+   Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
      console.log(result);
-     res.status(200).json({message: "Post Deleted"});
-   })
+     if(result.n > 0) {
+      res.status(200).json({message: "Post Deleted"});
+     } else {
+       res.status(401).json({message: "Not authorized"});
+     }
+   }).catch(error => {
+    res.status(500).json({
+      message: 'Fetching posts failed'
+    })
+  });
 });
 
 module.exports = router;
